eating is now working too :) seeding is in progress

+++ src/forth/forth.h
@@ -8,12 +8,12 @@
 #include "cstack.h"
 #include "dstack.h"
 
-typedef uint8_t identifier_t;
-typedef uint8_t generation_t;
+typedef uint32_t identifier_t;
+typedef uint16_t generation_t;
 typedef uint32_t step_t;
-typedef uint8_t coord_t;
+typedef uint16_t coord_t;
 typedef int8_t offset_t; /* [-1 .. 1] */
-typedef uint16_t energy_t;
+typedef uint32_t energy_t;
 
 
 typedef struct Forth {
